NewPlaylist FXML and Controller,+ changed setDuration to double since our variable is also double.Added a method to get all songs from the list as well a created and identifier for a playlist (im not sure if we gonna use id or name so i added both)

+++ MyTunes/src/be/Song.java
@@ -34,7 +34,7 @@ public class Song {
         this.songtitle = songtitle;
     }
 
-    public void setDuration(int duration) {
+    public void setDuration(double duration) {
         this.duration = duration;
     }
